Update prod pipeline to use new configuration
The Azure DevOps Pipeline Agents can't actually compile ARM64
versions of the dll and so files for Windows and Linux. They were
commented out until they are able to.

The macOS dylib can be built as a Universal library to support
both x64 and ARM64 architectures, so the build step was
modified to compile this library and copy it into the respective
folders for the architectures.